moved the component outside from apps
