+++ apps/web/app/page.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@repo/ui/button";
-import { Features } from "./components/features";
-import { HeroSection } from "./components/herosection";
-import { NavBar } from "./components/NavBar";
+import { Features } from "../components/features";
+import { HeroSection } from "../components/herosection";
+import { NavBar } from "../components/NavBar";
