+++ apps/web/components/dashboard.navbar.tsx
@@ -1,0 +1,19 @@
+
+
+export function DashboardNavbar() {
+
+    return <>
+        <div>
+            {/* logo */}
+            <div className="font-bold italic  ">
+                FarmVision
+            </div>
+            <div>
+                {/* profile */}
+                <div className="h-10 w-10 rounded-full bg-green-300 ">
+                    <p className="text-green-700">R</p>
+                </div>
+            </div>
+        </div>
+    </>
+}
