+++ apps/web/components/herosection.tsx
@@ -1,0 +1,30 @@
+
+
+
+export function HeroSection() {
+    return <>
+        <div>
+            <div>
+                {/* there will be a video that will disply here  */}
+            </div>
+
+            {/* this div will  be on top of the image or video use an image to get reference */}
+            <div>
+                {/* this will on the left side and on top of the Image  */}
+                <h1>
+                    EMPOWER YOUR FARM WITH AI Pick crops by market trends and connect with farmers—all in one.
+                </h1>
+            </div>
+
+            {/* to show the current weather */}
+            <div>
+                {/* this will on the right not big a small card that will show the weather  */}
+                {/* use html tag to show the weather  */}
+            </div>
+
+            <div>
+
+            </div>
+        </div>
+    </>
+}
