+++ apps/web/public/farm_Vision_Video.mp4:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=https://pixabay.com/
