+++ apps/web/tailwind.config.ts
@@ -4,7 +4,7 @@ import type { Config } from "tailwindcss";
 import sharedConfig from "@repo/tailwind-config";
 
 const config: Pick<Config, "content" | "presets"> = {
-  content: ["./app/**/*.tsx"],
+  content: ["./app/**/*.tsx", "./components/**/*.tsx"],
   presets: [sharedConfig],
 };
